Move stream_output to ProcessBuilder
